requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Executor
tensorflow
numpy
scipy
opencv-python
pillow
matplotlib
h5py
keras
https://github.com/OlafenwaMoses/ImageAI/releases/download/2.0.2/imageai-2.0.2-py3-none-any.whl
ImageHash
Flask-Migrate
flask-json-schema
psycopg2-binary
Flask-Login